+++ src/currency-format.filter.js
@@ -10,9 +10,11 @@ angular.module('currencyFormat', ['currencyFormat.iso'])
          * @param string currencyCode e.g. EUR, USD
          * @param number fractionSize User specified fraction size that overwrites default value
          * @param boolean useUniqSymbol use unique currency symbol
+         * @param string localeId e.g. ru_RU, default en_US
+         * @param boolean onlyAmount Retrieve amount only
          * @return string
          */
-        return function (amount, currencyCode, fractionSize = null, useUniqSymbol = true) {
+        return function (amount, currencyCode, fractionSize = null, useUniqSymbol = true, localeId = null, onlyAmount = false) {
             if (!currencyCode || Number(amount) != amount) {
                 return;
             }
@@ -32,17 +34,19 @@ angular.module('currencyFormat', ['currencyFormat.iso'])
 
             formatedAmount = formatedAmount.toFixed(currentFractionSize);
 
-            // Format numeral by language
+            // Format numeral by locale ID
 
-            var languageCode = $rootScope.currencyLanguage || 'en_US',
-                language = currencyFormatService.getLanguageByCode(languageCode);
+            localeId = localeId ? localeId : ($rootScope.currencyLanguage || 'en_US');
+            var languageOptions = currencyFormatService.getLanguageByCode(localeId);
 
-            formatedAmount = formatedAmount.split('.').join(language.decimal);
-            formatedAmount = formatedAmount.replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1' + language.thousands);
+            formatedAmount = formatedAmount.split('.').join(languageOptions.decimal);
+            formatedAmount = formatedAmount.replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1' + languageOptions.thousands);
 
             // Format currency
 
-            if (!!currency && !useUniqSymbol && !!currency.symbol && !!currency.symbol.template) {
+            if (onlyAmount) {
+                formattedCurrency = signAmount + formatedAmount;
+            } else if (!!currency && !useUniqSymbol && !!currency.symbol && !!currency.symbol.template) {
                 formattedCurrency = currency.symbol.template.replace('1', formatedAmount);
                 formattedCurrency = formattedCurrency.replace('$', currency.symbol.grapheme);
                 formattedCurrency = signAmount + formattedCurrency;